style: 컨벤션 통일
- 패키징 구조 맞추기
- 스웨게 응답 예시 추가
- 중복된 에러 처리 하나로 통합
